+++ lib/togls/feature_registry.rb
@@ -1,18 +1,22 @@
 module Togls
   class FeatureRegistry
-    def initialize
+    def initialize(base_rule_klass)
+      @base_rule_klass = base_rule_klass
       @registry = {}
-      @default_feature = Feature.new(:default, "the official default feature").on(Rule.new { false })
+      @default_feature = Feature.new(:default,
+                                     "the official default feature",
+                                     @base_rule_klass).on(Rule.new { false })
     end
 
-    def self.create(&features)
-      feature_registry = self.new
+    def self.create(base_rule_klass, &features)
+      feature_registry = self.new(base_rule_klass)
       feature_registry.instance_eval(&features) 
       feature_registry
     end
 
     def feature(tag, desc)
-      @registry[tag.to_sym] = Feature.new(tag.to_sym, desc)
+      @registry[tag.to_sym] = Feature.new(tag.to_sym, desc,
+                                          @base_rule_klass)
     end
 
     def get(key)