+++ lib/togls/rules.rb
@@ -1,4 +1,5 @@
 require 'togls/rules/boolean'
+require 'togls/rules/boolean_env_override'
 require 'togls/rules/group'
 
 module Togls