+++ spec/lib/togls/rules/boolean_spec.rb
@@ -12,7 +12,7 @@ describe Togls::Rules::Boolean do
   describe "#run" do
     it "returns the provided boolean value" do
       bool_rule = Togls::Rules::Boolean.new(true)
-      expect(bool_rule.run).to eq(true)
+      expect(bool_rule.run(double)).to eq(true)
     end
   end
 end