+++ spec/lib/togls/rules/group_spec.rb
@@ -13,7 +13,7 @@ describe Togls::Rules::Group do
     context "when the target is included in the list" do
       it "returns true" do
         group = Togls::Rules::Group.new(["target"])
-        expect(group.run("target")).to eq(true)
+        expect(group.run(double('feature_key'), "target")).to eq(true)
       end
     end
   end